requirements.txt
hikari
hikari-crescent
python-dotenv
uvloop
hikari-miru
aiosqlite